pyproject.toml
[project]
name = "wyoming-tts-proxy"
version = "0.1.0"
description = "Transparently normalize text and proxy requests to text-to-speech Wyoming services."
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "emoji>=2.15.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "wyoming>=1.8.0",
    "prometheus-client>=0.21.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]
